Добавила контроллер TripController
